pyproject.toml
[tool.poetry]
name = "habits-manager"
version = "0.1.0"
description = ""
authors = ["Maxim Donchuk <m.donchuk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2"
django-rest-framework = "^0.1.0"
django-environ = "^0.11.2"
celery = "^5.3.6"
drf-yasg = "^1.21.7"
psycopg2 = "^2.9.9"
setuptools = "^69.0.3"
djangorestframework-simplejwt = "^5.3.1"
ipython = "^8.19.0"
django-cors-headers = "^4.3.1"
requests = "^2.31.0"
django-celery-beat = "^2.5.0"
redis = "^5.0.1"
coverage = "^7.4.0"
flake8 = "^7.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
